Move Audio drivers from quantum to platform drivers folder
