+++ platforms/avr/drivers/audio_pwm_hardware.c
